+++ src/sabogaapi/routers/categories.py
@@ -3,7 +3,7 @@
 from fastapi import APIRouter, HTTPException
 
 from sabogaapi.logger import configure_logger
-from sabogaapi.schemas import Category
+from sabogaapi.schemas import Category, CategoryWithBoardgames
 from sabogaapi.services.category_service import CategoryService
 
 logger = configure_logger()
@@ -21,7 +21,7 @@ async def read_all_categories() -> list[Category]:
 
 
 @router.get("/{bgg_id}")
-async def read_category(bgg_id: int) -> Category:
+async def read_category(bgg_id: int) -> CategoryWithBoardgames:
     category = await CategoryService.read_category(bgg_id=bgg_id)
     if not category:
         raise HTTPException(status_code=404, detail="Category not found")