+++ src/sabogaapi/routers/designers.py
@@ -3,7 +3,7 @@
 from fastapi import APIRouter, HTTPException
 
 from sabogaapi.logger import configure_logger
-from sabogaapi.schemas import Designer, DesignerNetwork
+from sabogaapi.schemas import Designer, DesignerNetwork, DesignerWithBoardgames
 from sabogaapi.services import DesignerService
 
 logger = configure_logger()
@@ -26,7 +26,7 @@ async def read_designer_clusters() -> DesignerNetwork:
 
 
 @router.get("/{bgg_id}")
-async def read_designer(bgg_id: int) -> Designer:
+async def read_designer(bgg_id: int) -> DesignerWithBoardgames:
     designer = await DesignerService.read_designer(bgg_id=bgg_id)
     if not designer:
         raise HTTPException(status_code=404, detail="Designer not found")