+++ src/sabogaapi/scraper/_update.py
@@ -4,6 +4,7 @@ from pathlib import Path
 from typing import Any
 from zipfile import ZipFile
 
+import numpy as np
 import pandas as pd
 import requests
 from selenium import webdriver
@@ -50,6 +51,12 @@ def download_zip() -> pd.DataFrame:  # pragma: no cover
         time.sleep(2)
 
         cookie_button = driver.find_element(By.CLASS_NAME, "fc-cta-consent")
+
+        cookie_button.click()
+        logger.debug("Cookie consent clicked.")
+
+        cookie_button = driver.find_element(By.ID, "c-s-bn")
+
         cookie_button.click()
         logger.debug("Cookie consent clicked.")
 
@@ -57,7 +64,7 @@ def download_zip() -> pd.DataFrame:  # pragma: no cover
         password = settings.bgg_password
 
         driver.find_element(By.ID, "inputUsername").send_keys(username)
-        driver.find_element(By.NAME, "password").send_keys(password)
+        driver.find_element(By.ID, "inputPassword").send_keys(password)
         driver.find_element(By.CSS_SELECTOR, "[type='submit']").click()
 
         logger.info("Login submitted. Waiting for redirect")
@@ -144,24 +151,56 @@ async def insert_games(games_df: pd.DataFrame) -> tuple[list[Any], int]:
         rank_history = await models.RankHistory.find(
             models.RankHistory.bgg_id == game.bgg_id,
         ).to_list()
+        for doc in rank_history:
+            new_date = datetime.datetime(
+                doc.date.year, doc.date.month, doc.date.day, tzinfo=datetime.UTC
+            )
+            new_date = doc.date.replace(hour=0, minute=0, second=0, microsecond=0)
+            if doc.date != new_date:
+                await doc.delete()
+                corrected = models.RankHistory(
+                    date=new_date,
+                    bgg_id=doc.bgg_id,
+                    bgg_rank=doc.bgg_rank,
+                    bgg_geek_rating=doc.bgg_geek_rating,
+                    bgg_average_rating=doc.bgg_average_rating,
+                )
+                await corrected.insert()
+
         rank_volatility, geek_rating_volatility, average_rating_volatility = (
             calculate_volatility(
                 [schemas.RankHistory(**entry.model_dump()) for entry in rank_history],
             )
         )
-        game.bgg_rank_volatility = rank_volatility
-        game.bgg_geek_rating_volatility = geek_rating_volatility
-        game.bgg_average_rating_volatility = average_rating_volatility
+        game.bgg_rank_volatility = (
+            np.nan_to_num(rank_volatility) if rank_volatility is not None else 0
+        )
+        game.bgg_geek_rating_volatility = (
+            np.nan_to_num(geek_rating_volatility)
+            if geek_rating_volatility is not None
+            else 0
+        )
+        game.bgg_average_rating_volatility = (
+            np.nan_to_num(average_rating_volatility)
+            if average_rating_volatility is not None
+            else 0
+        )
 
         rank_trend, geek_rating_trend, average_rating_trend, mean_trend = (
             calculate_trends(
                 [schemas.RankHistory(**entry.model_dump()) for entry in rank_history]
             )
         )
-        game.bgg_rank_trend = rank_trend
-        game.bgg_geek_rating_trend = geek_rating_trend
-        game.bgg_average_rating_trend = average_rating_trend
-        game.mean_trend = mean_trend
+        game.bgg_rank_trend = np.nan_to_num(rank_trend) if rank_trend is not None else 0
+        game.bgg_geek_rating_trend = (
+            np.nan_to_num(geek_rating_trend) if geek_rating_trend is not None else 0
+        )
+        game.bgg_average_rating_trend = (
+            np.nan_to_num(average_rating_trend)
+            if average_rating_trend is not None
+            else 0
+        )
+        game.mean_trend = np.nan_to_num(mean_trend) if mean_trend is not None else 0
 
         await game.save()