+++ src/sabogaapi/services/__init__.py
@@ -1,5 +1,17 @@
 from .boardgame_service import BoardgameService
+from .category_service import CategoryService
 from .designer_service import DesignerService
+from .family_service import FamilyService
+from .mechanic_service import MechanicService
 from .rank_history_service import RankHistoryService
+from .search_service import SearchService
 
-__all__ = ["BoardgameService", "DesignerService", "RankHistoryService"]
+__all__ = [
+    "BoardgameService",
+    "CategoryService",
+    "DesignerService",
+    "FamilyService",
+    "MechanicService",
+    "RankHistoryService",
+    "SearchService",
+]